Add dbName to MariaDB Module
This commit adds a new paramater of dbName to the MariaDB module that
allows you to specify a database name that you can use for your own
purposes. If one is not provided we start the container with a
'sample-database'

This also splits out the Serve() and Base() function to allow you to
grab the dagger.Container instead of the dagger.Service if needed.

Lastly, we add a Debug() function that allows us to connect to the
service and actually test it out if needed. Note the interface here is a
bit strange because from the Dagger Perspective Serve() accepts optional
arguments, but because I am calling the function directly from Go, these
are no longer optional.