Release v0.2.0 of Docusaurus module
This commit splits out the build and serve portion of this module so
that it makes it easier to debug your site if needed. We also renamed
run to serve to make it more obvious what is going on. The end result is
a simplified module that does not require any chaining in order to work.