Release v0.4.0 of Docusaurus
This fixes a bug with disabling cache logic, it also allows you to
specify a custom volume name to take care of being able to change
the name of the volume without needing to completely disable the cache.

We also implement the `Build` method that builds the production version
of your docs and allows you to pass them around.

Lastly, we separate Serve and ServeDev to allow you to easily serve both
the production documentation and the development version.