remove version string from API
I was getting a "426 upgrade required" error message when trying to use
this module. The docs do not make it clear which version of the API I am
supposed to be using. So I tried to call it without specifying the
version at all and this seemed to work just fine.

Signed-off-by: Lev Lazinskiy <lev@levlaz.org>